Imagen

+++ src/model/Entidad.cpp
@@ -1,6 +1,6 @@
 #include "Entidad.h"
 
-Entidad::Entidad(int posX, int posY, int ancho, int alto) {
+Entidad::Entidad(float posX, float posY, int ancho, int alto) {
     this->posX = posX;
     this->posY = posY;
     this->ancho = ancho;
